requirements.txt
openai>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
trafilatura>=1.6.0
lxml>=5.0.0
playwright>=1.40.0
streamlit>=1.35.0
pandas>=2.1.0
